requirements.txt
aniso8601==8.1.1
asgiref==3.3.4
autopep8==1.5.6
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2
django-cors-headers==3.7.0
django-graphql-jwt==0.3.2
django-sslserver==0.22
ecdsa==0.14.1
graphene==3.0b7
graphene-django==3.0.0b7
graphql-core==3.1.4
graphql-relay==3.1.0
idna==2.10
oauthlib==3.1.0
pkg-resources==0.0.0
promise==2.3
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.0.1
python-dotenv==0.17.0
python-jose==3.2.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
Rx==1.6.1
singledispatch==3.6.1
six==1.15.0
social-auth-app-django==3.4.0
social-auth-core==4.1.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.4
